pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "en-core-web-sm",
    "matplotlib>=3.10.8",
    "numpy<2.0.0",
    "pandas<3.0.0",
    "protobuf>=6.33.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.3",
    "sentencepiece>=0.2.1",
    "spacy<3.8.0",
    "tiktoken>=0.12.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl" }
